requirements.txt
flask
flask_cors
pytesseract
opencv-python
Pillow
flask
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.2.0
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
setuptools==68.2.2

requests
cython